Update gyro.h
Follow your own naming conventions please.

AcX is not very descriptive. The code when complied doesn't care about the 'name' or its length, just cares about what it does. So make it easier on yourself and name things better.

From looking at the implementation of the function 'FunctionsPitchRoll' I would assume the A,B and C inputs are the axis it is trying to calibrate too. Naming them such would probably help.